% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burden_Effect.R
\name{Burden_Effect}
\alias{Burden_Effect}
\title{Calculate Burden Effect Size}
\usage{
Burden_Effect(
  genotype,
  obj_nullmodel,
  beta_par = c(1, 1),
  annotation_phred = NULL,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2
)
}
\arguments{
\item{genotype}{A matrix or dgCMatrix of genotypes.}

\item{obj_nullmodel}{A null model object containing information about the model fit and residuals.}

\item{beta_par}{A vector of length 2 specifying the parameters of the beta distribution used for weighting. Default is \code{c(1, 1)}.}

\item{annotation_phred}{A numeric vector of annotation Phred scores. Default is \code{NULL}.}

\item{rare_maf_cutoff}{The cutoff for maximum minor allele frequency to define rare variants. Default is \code{0.01}.}

\item{rv_num_cutoff}{The minimum number of variants required in the set for analysis. Default is \code{2}.}
}
\value{
A list containing:
\item{num_variant}{The number of rare variants in the set.}
\item{cMAC}{The cumulative minor allele count of the variants.}
\item{RV_label}{A vector indicating which variants are rare.}
\item{Burden_Effect_Size}{A vector containing the burden effect size, standard error, score statistic, score standard error, and p-value.}
\item{Burden_Variable}{The burden variable weighted by the annotation score for each individual.}
}
\description{
This function calculates the burden effect size for a given genotype matrix, considering relatedness of individuals and annotations.
}
