% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Centric_Coding_Burden_each.R
\name{Gene_Centric_Coding_Burden_each}
\alias{Gene_Centric_Coding_Burden_each}
\title{Get the Burden variable and Burden effect size for each gene}
\usage{
Gene_Centric_Coding_Burden_each(
  chr,
  gene_name,
  category = c("plof", "plof_ds", "missense", "disruptive_missense", "synonymous"),
  genofile,
  obj_nullmodel,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  QC_label = "annotation/filter",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Beta_Annotation = "Burden(1,1)",
  silent = FALSE
)
}
\arguments{
\item{chr}{chromosome.}

\item{gene_name}{name of the gene that is significant in the existing summary statistics.}

\item{category}{the coding functional category of \code{gene_name} that is significant in STAAR procedure. Choices include
\code{plof}, \code{plof_ds}, \code{missense}, \code{disruptive_missense}, \code{synonymous} (default = \code{all_categories}).}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "SNV", "Indel", or "variant" (default = "SNV").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file \cr (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the name and the corresponding channel name in the aGDS file.}

\item{silent}{logical: should the report of error messages be suppressed (default = FALSE).}
}
\value{
a list containing the information of the given gene, Burden variable, and Burden effect size.
}
\description{
The \code{Gene_Centric_Coding_Burden_each} function takes in chromosome, gene name, functional category,
the object of opened annotated GDS file, and the object from fitting the null model to get the Burden variables and Burden Effect Size.
}
