% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individual_analysis_19}
\alias{individual_analysis_19}
\title{Individual Analysis Data}
\format{
Individual variant association results from STAARpipeline.
}
\source{
The data is provided as part of the `mr.carv` package.
}
\usage{
individual_analysis_19
}
\description{
This dataset is the individual variant association with PTB stored in a text file `individual_analysis_19.out`.
}
\examples{
# Load the sample data from the package
data(individual_analysis_19)
str(individual_analysis_19)
}
\keyword{datasets}
