% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Estimate.R
\name{Individual_Estimate}
\alias{Individual_Estimate}
\title{Estimate Individual Variant Effects}
\usage{
Individual_Estimate(
  chr,
  df_indv,
  genofile,
  obj_nullmodel,
  mac_cutoff = 20,
  QC_label = "annotation/filter",
  variant_type = c("variant", "SNV", "Indel"),
  geno_missing_imputation = c("mean", "minor")
)
}
\arguments{
\item{chr}{Chromosome number.}

\item{df_indv}{A data frame containing individual variant information with columns: `CHR`, `POS`, `annotation.id`, `REF`, and `ALT`.}

\item{genofile}{An opened annotated GDS (aGDS) file object containing genotype data.}

\item{obj_nullmodel}{A null model object containing information about the model fit and residuals.}

\item{mac_cutoff}{Minor allele count cutoff for rare variants. Default is 20.}

\item{QC_label}{Quality control label. Default is `"annotation/filter"`.}

\item{variant_type}{Type of variant to include in the analysis. Choices are `"variant"`, `"SNV"`, or `"Indel"`. Default is `"variant"`.}

\item{geno_missing_imputation}{Method for handling missing genotypes. Choices are `"mean"` or `"minor"`. Default is `"mean"`.}
}
\value{
A list containing:
\item{Variant_Estimate}{A data frame with columns: `CHR`, `variant.id`, `annotation.id`, `POS`, `REF`, `ALT`, `ALT_AF`, `MAF`, `N`, `pvalue`, `pvalue_log10`, `Score`, `Score_se`, `Est`, `Est_se`.}
\item{Variant_dosg}{A matrix of genotype dosages for the selected variants.}
}
\description{
This function estimates the effects of individual variants on the outcome, given the variant information, genotype data, and a null model.
}
