% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_LE_Estimate.R
\name{select_LE_Estimate}
\alias{select_LE_Estimate}
\title{Select the genes and single variant that is in linkage equilibrium}
\usage{
select_LE_Estimate(
  chr,
  df_indv,
  df_gene,
  genofile,
  obj_nullmodel,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  QC_label = "annotation/filter",
  variant_type_indv = c("SNV", "Indel", "variant"),
  variant_type_gene = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  le_threshold = 0.1,
  pvalues_weight = FALSE,
  silent = FALSE
)
}
\arguments{
\item{chr}{Chromosome number.}

\item{df_indv}{Dataframe containing individual variant information with columns: CHR, POS, annotation.id, REF, ALT, Est, Est_se, and pvalue.}

\item{df_gene}{Dataframe containing gene information with columns: CHR, gene_name, category, annotation, Est, Est_se, and pvalue.}

\item{genofile}{An object of an opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{An object from fitting the null model, which is either the output from the \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{rare_maf_cutoff}{Minor allele frequency cutoff for rare variants. Default is 0.01.}

\item{rv_num_cutoff}{Cutoff for the minimum number of variants in a given variant-set. Default is 2.}

\item{QC_label}{Channel name of the QC label in the GDS/aGDS file. Default is "annotation/filter".}

\item{variant_type_indv}{Type of individual variants to include in the analysis. Choices include "SNV", "Indel", or "variant". Default is c("SNV", "Indel", "variant").}

\item{variant_type_gene}{Type of gene-centric variants to include in the analysis. Choices include "SNV", "Indel", or "variant". Default is c("SNV", "Indel", "variant").}

\item{geno_missing_imputation}{Method for imputing missing genotypes. Choices are "mean" or "minor". Default is c("mean", "minor").}

\item{Annotation_dir}{Directory for functional annotation information in the aGDS file. Default is "annotation/info/FunctionalAnnotation".}

\item{Annotation_name_catalog}{A data frame containing the name and the corresponding channel name in the aGDS file.}

\item{le_threshold}{Threshold for linkage disequilibrium (LD) correlation to define independent sets. Default is 0.1.}

\item{pvalues_weight}{Logical; if TRUE, uses p-values to weight the selection of independent SNPs, if FALSE, randomly select one set with independent SNPs. Default is FALSE.}

\item{silent}{Logical; if TRUE, suppresses messages. Default is FALSE.}
}
\value{
A list containing:
\describe{
  \item{X_indv}{Dataframe of individual variants with selected independent SNPs for exposure.}
  \item{X_gene}{Dataframe of gene-centric variants with selected independent SNPs for exposure.}
  \item{Y_indv}{Dataframe of effect size estimates for individual variants for outcome.}
  \item{Y_gene}{Dataframe of effect size estimates for gene-centric variants for outcome.}
}
}
\description{
The \code{select_LE_Estimate} function find the individual variants and gene-centric variants that are in linakge equilibrium
on a given chromosome. It calculates the correlation between genes and individual SNPs and selects a maximal independent set
of SNPs.
}
