% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC_CHR19}
\alias{TC_CHR19}
\title{TC_CHR19 Data}
\format{
A list with two data frames:
\describe{
  \item{indv_effect}{A data frame with columns:
    \itemize{
      \item{CHR}{Numeric, chromosome}
      \item{POS}{Numeric, the position of each individual variant}
      \item{REF}{Character, reference allele of each individual variant}
      \item{ALT}{Character, alternative allele of each individual variant}
      \item{annotation.id}{Character, the annotation id of each individual variant}
      \item{Est}{Numeric, the effect size estimates of each individual variant on the exposure total cholesterol}
      \item{Est_se}{Numeric, the standard error of the effect size estimates for each individual variant on the exposure total cholesterol}
      \item{pvalue}{Numeric, the P value of the effect size for each individual variant}
      \item{MAF}{Numeric, optional, for the harmonization}
    }
  }
  \item{gene_effect}{A data frame with columns:
    \itemize{
      \item{gene_name}{Character, gene name}
      \item{CHR}{Numeric, chromosome}
      \item{annotation}{Character, the Burden weight}
      \item{category}{Character, the coding functional category used in the results of STAARpipeline}
      \item{Est}{Numeric, the Burden effect size estimates of gene-centric coding rare variant aggregate sets}
      \item{Est_se}{Numeric, the standard error of Burden effect size estimates of gene-centric coding rare variant aggregate sets}
      \item{pvalue}{Numeric, the P value of the Burden effect size estimates of gene-centric coding rare variant aggregate sets}
    }
  }
}
}
\source{
The data is provided as part of the `mr.carv` package.
}
\usage{
TC_CHR19
}
\description{
This dataset is stored in an Excel file `TC_CHR19.xlsx` which contains two sheets:
- `indv_effect`: Individual variant associations with total cholesterol
- `gene_effect`: Gene-Centric coding associations with total cholesterol
}
\examples{
# Load the sample data from the package
data(TC_CHR19)
str(TC_CHR19$indv_effect)
str(TC_CHR19$gene_effect)
}
\keyword{datasets}
