% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Window_Burden_each.R
\name{Window_Burden_each}
\alias{Window_Burden_each}
\title{Get the Burden variable and Burden effect size for each sliding/dynamic window region}
\usage{
Window_Burden_each(
  chr,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  QC_label = "annotation/filter",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Beta_Annotation = "Burden(1,1)",
  silent = FALSE
)
}
\arguments{
\item{chr}{chromosome.}

\item{start_loc}{the start location of the sliding/dynamic window region that is significant in the existing summary statistics.}

\item{end_loc}{the end location of the sliding/dynamic window region that is significant in the existing summary statistics.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "SNV", "Indel", or "variant" (default = "SNV").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file \cr (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the name and the corresponding channel name in the aGDS file.}

\item{Beta_Annotation}{annotation weight used in the summary statistics.}

\item{silent}{logical: should the report of error messages be suppressed (default = FALSE).}
}
\value{
a list containing the information of the given gene, Burden variable, and Burden effect size.
}
\description{
The \code{Window_Burden_each} function takes in chromosome, start location, end location,
the object of opened annotated GDS file, and the object from fitting the null model to get the Burden variables and Burden Effect Size.
}
